Fix apis to include authorization header
